requirements.txt
sqlitedict
tkmacosx
pillow